start moving version info from init to _version.py
